custom_delimiters -> custom_syntax and restructure types
