+++ minijinja/Cargo.toml
@@ -13,7 +13,7 @@ rust-version = "1.61"
 exclude = ["tests"]
 
 [package.metadata.docs.rs]
-features = ["source", "json", "urlencode", "testutils", "custom_delimiters"]
+features = ["source", "json", "urlencode", "testutils", "custom_syntax"]
 rustdoc-args = ["--cfg", "docsrs", "--html-in-header", "doc-header.html"]
 
 [features]
@@ -26,7 +26,7 @@ debug = []
 source = ["self_cell", "memo-map"]
 unicode = ["unicode-ident"]
 testutils = []
-custom_delimiters = ["dep:aho-corasick"]
+custom_syntax = ["dep:aho-corasick"]
 
 # Speedups
 key_interning = []
