+++ minijinja/src/compiler/parser.rs
@@ -1,12 +1,11 @@
 use std::collections::BTreeSet;
 use std::fmt;
-use std::sync::Arc;
 
 use crate::compiler::ast::{self, Spanned};
 use crate::compiler::lexer::tokenize;
 use crate::compiler::tokens::{Span, Token};
+use crate::custom_syntax::SyntaxConfig;
 use crate::error::{Error, ErrorKind};
-use crate::settings::SyntaxConfig;
 use crate::value::Value;
 
 const MAX_RECURSION: usize = 150;
@@ -97,7 +96,7 @@ struct TokenStream<'a> {
 
 impl<'a> TokenStream<'a> {
     /// Tokenize a template
-    pub fn new(source: &'a str, in_expr: bool, syntax: Arc<SyntaxConfig>) -> TokenStream<'a> {
+    pub fn new(source: &'a str, in_expr: bool, syntax: SyntaxConfig) -> TokenStream<'a> {
         let mut iter = Box::new(tokenize(source, in_expr, syntax)) as Box<dyn Iterator<Item = _>>;
         let current = iter.next();
         TokenStream {
@@ -218,7 +217,7 @@ macro_rules! with_recursion_guard {
 }
 
 impl<'a> Parser<'a> {
-    pub fn new(source: &'a str, in_expr: bool, syntax: Arc<SyntaxConfig>) -> Parser<'a> {
+    pub fn new(source: &'a str, in_expr: bool, syntax: SyntaxConfig) -> Parser<'a> {
         Parser {
             stream: TokenStream::new(source, in_expr, syntax),
             in_macro: false,
@@ -1102,7 +1101,7 @@ pub fn parse<'source>(source: &'source str, filename: &str) -> Result<ast::Stmt<
 pub fn parse_with_syntax<'source>(
     source: &'source str,
     filename: &str,
-    syntax: Arc<SyntaxConfig>,
+    syntax: SyntaxConfig,
 ) -> Result<ast::Stmt<'source>, Error> {
     // we want to chop off a single newline at the end.  This means that a template
     // by default does not end in a newline which is a useful property to allow
@@ -1126,7 +1125,7 @@ pub fn parse_with_syntax<'source>(
 }
 
 /// Parses an expression
-pub fn parse_expr(source: &str, syntax: Arc<SyntaxConfig>) -> Result<ast::Expr<'_>, Error> {
+pub fn parse_expr(source: &str, syntax: SyntaxConfig) -> Result<ast::Expr<'_>, Error> {
     let mut parser = Parser::new(source, true, syntax);
     parser
         .parse_expr()
