+++ minijinja/src/error.rs
@@ -137,7 +137,7 @@ pub enum ErrorKind {
     /// Engine ran out of fuel
     #[cfg(feature = "fuel")]
     OutOfFuel,
-    #[cfg(feature = "custom_delimiters")]
+    #[cfg(feature = "custom_syntax")]
     /// Error creating aho-corasick delimiters
     InvalidDelimiter,
 }
@@ -165,7 +165,7 @@ impl ErrorKind {
             ErrorKind::WriteFailure => "failed to write output",
             #[cfg(feature = "fuel")]
             ErrorKind::OutOfFuel => "engine ran out of fuel",
-            #[cfg(feature = "custom_delimiters")]
+            #[cfg(feature = "custom_syntax")]
             ErrorKind::InvalidDelimiter => "invalid custom delimiters",
         }
     }
