+++ minijinja/src/lib.rs
@@ -173,7 +173,7 @@
 //!   the previous uses of key interning are no longer needed.  Enabling it however
 //!   cuts down on memory usage slightly in certain scenarios by interning all string
 //!   keys used in dynamic map values.
-//! - `custom_delimiters`: when this feature is enabled, custom delimiters are supported by
+//! - `custom_syntax`: when this feature is enabled, custom delimiters are supported by
 //!   the parser.
 //!
 //! </details>
@@ -187,13 +187,13 @@
 mod macros;
 
 mod compiler;
+mod custom_syntax;
 mod defaults;
 mod environment;
 mod error;
 mod expression;
 mod key;
 mod output;
-mod settings;
 mod template;
 mod utils;
 mod vm;
@@ -222,8 +222,8 @@ pub use self::output::Output;
 pub use self::template::Template;
 pub use self::utils::{AutoEscape, HtmlEscape, UndefinedBehavior};
 
-#[cfg(feature = "custom_delimiters")]
-pub use self::settings::Syntax;
+#[cfg(feature = "custom_syntax")]
+pub use self::custom_syntax::Syntax;
 
 #[cfg(feature = "source")]
 pub use self::source::Source;
@@ -246,7 +246,7 @@ pub mod machinery {
     pub use crate::compiler::lexer::tokenize;
     pub use crate::compiler::parser::{parse, parse_with_syntax};
     pub use crate::compiler::tokens::{Span, Token};
-    pub use crate::settings::SyntaxConfig;
+    pub use crate::custom_syntax::SyntaxConfig;
     pub use crate::template::CompiledTemplate;
     pub use crate::vm::Vm;
